feat(nextjs): introduce dedicated /pages entry for Pages Router support
