+++ apps/nextjs/pages-router/package.json
@@ -11,7 +11,7 @@
     "format:check": "oxfmt --check"
   },
   "dependencies": {
-    "@aura-stack/react": "workspace:*",
+    "@aura-stack/next": "workspace:*",
     "@radix-ui/react-slot": "^1.2.4",
     "lucide-react": "catalog:lucide-react",
     "next": "catalog:next",
