+++ apps/nextjs/pages-router/src/contexts/auth.tsx
@@ -1,4 +1,4 @@
-import { AuthProvider as AuraAuthProvider, type AuthProviderProps } from "@aura-stack/react"
+import { AuthProvider as AuraAuthProvider, type AuthProviderProps } from "@aura-stack/next/client"
 import { authClient } from "@/lib/auth-client"
 
 export const AuthProvider = ({ children, initialSession }: Omit<AuthProviderProps, "client">) => {
