+++ apps/nextjs/pages-router/src/lib/auth-client.ts
@@ -1,4 +1,4 @@
-import { createAuthClient } from "@aura-stack/react"
+import { createAuthClient } from "@aura-stack/next/client"
 
 export const authClient = createAuthClient({
     baseURL: "http://localhost:3000",
