+++ apps/nextjs/pages-router/src/lib/auth.ts
@@ -1,10 +1,10 @@
-import { createAuth } from "@aura-stack/react/server"
-import { createSecretValue } from "@aura-stack/react/crypto"
-import { builtInOAuthProviders, type BuiltInOAuthProvider } from "@aura-stack/react/oauth"
+import { createAuth } from "@aura-stack/next/pages"
+import { createSecretValue } from "@aura-stack/next/crypto"
+import { builtInOAuthProviders, type BuiltInOAuthProvider } from "@aura-stack/next/oauth"
 
 export const oauth = Object.keys(builtInOAuthProviders) as BuiltInOAuthProvider[]
 
-export const { api, handlers } = createAuth({
+export const { api, handlers, toHandler } = createAuth({
     oauth,
     basePath: "/api/auth",
     baseURL: "http://localhost:3000",
