+++ packages/next/CHANGELOG.md
@@ -10,6 +10,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
+- Introduced the `createAuth` function to manage the Next.js Pages Router. It includes the `toHandler` adapter for handling Node.js requests and is accessible from the `/pages` entry module. [#169](https://github.com/aura-stack-ts/auth/pull/169)
+
 - Updated Next.js API functions (`signIn`, `signInCredentials`, `updateSession`, and `signOut`) to align with the standardized core API contracts, including improved conditional return types for redirect flows, consistent cookie synchronization from API response headers, and support for redirect-aware `updateSession` behavior in App Router helpers. [#146](https://github.com/aura-stack-ts/auth/pull/146)
 
 - Introduced a seamless Next.js integration package that encapsulates Next.js authentication logic for both client-side and server-side rendering by integrating the `@aura-stack/react` package. It supports both rendering strategies, with primary support for server-side rendering (SSR). For client-side usage, import the modules from the `/client` entry point. [#141](https://github.com/aura-stack-ts/auth/pull/141)
