+++ packages/next/deno.json
@@ -9,7 +9,8 @@
     "./identity": "./src/_core/identity.ts",
     "./crypto": "./src/_core/crypto.ts",
     "./shared": "./src/_core/shared.ts",
-    "./types": "./src/@types/index.ts"
+    "./types": "./src/@types/index.ts",
+    "./pages": "./src/pages/index.ts"
   },
   "imports": {
     "@/": "./src/",
