+++ packages/next/src/lib/utils.ts
@@ -1,0 +1,10 @@
+export const isPagesRouter = () => {
+    try {
+        ;(async () => {
+            await import("next/headers")
+        })()
+        return true
+    } catch {
+        return false
+    }
+}
