+++ packages/next/src/pages/index.ts
@@ -1,0 +1,2 @@
+export { createAuth } from "@/pages/createAuth"
+export type { User, Session, AuthConfig, AuthInstance } from "@/@types/core.ts"
