+++ packages/next/tsdown.config.ts
@@ -13,6 +13,7 @@ export default defineConfig({
         "src/_core/crypto.ts",
         "src/_core/shared.ts",
         "src/@types/index.ts",
+        "src/pages/index.ts",
     ],
     fixedExtension: false,
     deps: {
